Work on SurveySubmissionCard (WIP)
