+++ front/app/containers/UsersShowPage/Surveys/SurveySubmissionCard/Body.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Body = () => {
+  return <></>;
+};
+
+export default Body;
